requirements.txt
asn1crypto==0.24.0
Babel==2.6.0
backcall==0.1.0
bleach==3.0.2
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
colorama==0.4.0
cryptography==2.4.2
decorator==4.3.0
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.1.4
django-crispy-forms==1.7.2
django-import-export==1.1.0
django-phonenumber-field==2.1.0
djangorestframework==3.7.7
docopt==0.6.2
entrypoints==0.2.3
et-xmlfile==1.0.1
idna==2.7
ipykernel==5.1.0
ipython==7.1.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
jdcal==1.4
jedi==0.13.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
MarkupSafe==1.1.0
mistune==0.8.4
nbconvert==5.4.0
nbformat==4.4.0
notebook==5.7.0
numpy==1.15.2
odfpy==1.4.0
openpyxl==2.5.12
packaging==18.0
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
phonenumbers==8.10.2
pickleshare==0.7.5
Pillow==5.3.0
pipdeptree==0.13.1
prometheus-client==0.4.2
prompt-toolkit==2.0.7
pycparser==2.19
pygeoip==0.3.2
Pygments==2.2.0
pyOpenSSL==18.0.0
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.7
pywinpty==0.5.4
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.3
requests==2.20.0
semantic-version==2.6.0
Send2Trash==1.5.0
six==1.11.0
tablib==0.12.1
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
unicodecsv==0.14.1
updatable==0.3.1
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
xlrd==1.2.0
xlwt==1.3.0
yarg==0.1.9